支払い情報とチームの状態取得はPaymentProviderに移動し、PaymentProviderはPrivateRouteの中で読み込む
Homeで実行しているとHomeコンポーネントが描画される度に支払い情報とチームの状態取得の処理が走ってしまうため
